# perf: faster text-table