--- a/lib/plugins/console/list/category.ts
+++ b/lib/plugins/console/list/category.ts
@@ -1,5 +1,5 @@
 import { underline } from 'picocolors';
-import table from 'text-table';
+import table from 'fast-text-table';
 import { stringLength } from './common';
 import type Hexo from '../../../hexo';
 import type { CategorySchema } from '../../../types';