--- a/lib/plugins/console/list/page.ts
+++ b/lib/plugins/console/list/page.ts
@@ -1,5 +1,5 @@
 import { magenta, underline, gray } from 'picocolors';
-import table from 'text-table';
+import table from 'fast-text-table';
 import { stringLength } from './common';
 import type Hexo from '../../../hexo';
 import type { PageSchema } from '../../../types';