--- a/lib/plugins/console/list/tag.ts
+++ b/lib/plugins/console/list/tag.ts
@@ -1,5 +1,5 @@
 import { magenta, underline } from 'picocolors';
-import table from 'text-table';
+import table from 'fast-text-table';
 import { stringLength } from './common';
 import type Hexo from '../../../hexo';
 import type { TagSchema } from '../../../types';