--- a/package.json
+++ b/package.json
@@ -42,6 +42,7 @@
     "abbrev": "^3.0.0",
     "archy": "^1.0.0",
     "bluebird": "^3.7.2",
+    "fast-text-table": "^1.0.1",
     "hexo-cli": "^4.3.2",
     "hexo-front-matter": "^4.2.1",
     "hexo-fs": "^5.0.0",
@@ -58,7 +59,6 @@
     "picocolors": "^1.1.1",
     "pretty-hrtime": "^1.0.3",
     "strip-ansi": "^6.0.0",
-    "text-table": "^0.2.0",
     "tildify": "^2.0.0",
     "titlecase": "^1.1.3",
     "warehouse": "^6.0.0"
@@ -74,7 +74,6 @@
     "@types/node": "^20.17.6",
     "@types/nunjucks": "^3.2.2",
     "@types/sinon": "^17.0.3",
-    "@types/text-table": "^0.2.4",
     "0x": "^5.1.2",
     "c8": "^9.0.0",
     "chai": "^4.3.6",